+++ krillcommand.py
@@ -5,7 +5,7 @@ def getKrillMessage(userID:str):
     dateNow = str(datetime.today().strftime('%m/%d'))
     integer = 0
     message = ''
-    if not dateNow == None:message = 'Krilled ' + userID
+    message = 'Krilled ' + userID
     if dateNow == '04/01':integer = random.randrange(0,9,1); print('the random int is' + str(integer))
     if dateNow == '04/01' and integer < 3:message = 'Not so shrimple now is it' + userID + '?'
     if dateNow == '04/01' and integer < 6 and integer > 3:message = 'Krilled' + userID + '\n its as shrimple as that.'